+++ src/components/Main/Pagination/type.ts
