+++ src/components/Main/PostCard/PostSnippet/type.tsx
@@ -1,0 +1,8 @@
+export interface PostSnippetProps {
+  avatar: string;
+  image?: string;
+  title: string;
+  count: string;
+  fullName?: string;
+  userId?: string;
+}
