+++ src/components/Main/PostCard/type.ts
@@ -11,12 +11,3 @@ export interface PostSnippet {
 export interface PostCardProps {
   post: PostSnippet;
 }
-
-export interface PostSnippetProps {
-  avatar: string;
-  image?: string;
-  title: string;
-  count: string;
-  fullName?: string;
-  userId?: string;
-}
