+++ src/components/Main/UserListCard/UserSnippet/type.ts
