+++ src/components/Main/style.tsx
@@ -1,0 +1,12 @@
+import styled from 'styled-components';
+
+export const UserSnippetBox = styled.div`
+  display: flex;
+  align-items: center;
+  gap: 10px;
+  position: relative;
+
+  margin: 4px 0;
+
+  cursor: pointer;
+`;
